+++ lib/swid/frontend/dart/extractStaticConstFromSyntacticEntity.dart
